requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
aiosqlite==0.19.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
slowapi==0.1.9
psycopg2-binary==2.9.9
alembic==1.13.1
beautifulsoup4==4.12.3
reportlab==4.1.0
